refactor(academy): split academy page into per-route slices (AYC-000)
