+++ ayunis-core-frontend/src/pages/SUMMARY.md
@@ -4,4 +4,4 @@ Page components composing widgets and features for each application route.
 
 Route-level page modules handle authentication, chat conversations, agent management, prompt libraries, knowledge base management, skill management, user settings, organization admin settings, and super-admin platform management. Each page follows a consistent internal structure with api/, ui/, and model/ subdirectories separating data hooks, components, and types.
 
-The `auth/` group contains the login, register, email confirmation, password forgot/reset, and invite acceptance flows. The `chat/` and `new-chat/` pages manage AI conversation interfaces with message streaming, agent selection, knowledge base attachment (`useKnowledgeBaseAttachment` hook for adding/removing KBs on existing threads, plus pending-KB context state for new threads), and side-panel artifact editing flows such as changing an artifact's assigned letterhead. The `chats/` page lists conversation threads. The `agents/` and `agent/` pages handle AI assistant browsing and detailed configuration. The `prompts/` page provides prompt library management. The `knowledge-bases/` page lists knowledge bases with creation and deletion, while `knowledge-base/` provides the detail view for a single knowledge base including document upload, removal, and property editing. The `skills/` page lists available skills, and `skill/` shows skill details. The `academy/` page module presents onboarding learning content: `AcademyPage` lists chapters and `ChapterDetailPage` shows a single chapter with an embedded Loom video, using the `toLoomEmbedUrl` helper in `lib/` to rewrite Loom share URLs into embeddable iframe URLs; `ChapterQuizPage` runs the chapter quiz and, once the whole academy is completed, both the `QuizResultCard` and the completion banner on `AcademyPage` offer the KI-Führerschein certificate PDF download via the shared `useDownloadCertificate` hook from `features/academy`. The `install/` page handles app installation flows. User `settings/` includes account, general preferences, and `chat-settings/` for per-user AI conversation configuration (custom system prompts). The `admin-settings/` group covers organization administration: model configuration, MCP integrations, team/user management, security, letterhead management (list in `letterheads-settings/`, detail/edit in `letterhead-detail/`), and usage analytics. The `super-admin-settings/` group provides platform-wide management for the models catalog, organization oversight, and platform configuration including per-tier fair-use message budgets via `FairUseLimitsSection` (with `useFairUseLimits` and `useSetFairUseLimit` hooks performing `windowMs ↔ windowHours` conversion at the hook boundary). Each page uses the hooks pattern—one custom hook per operation (create, update, delete)—encapsulating TanStack Query mutations with toast notifications and cache invalidation.
+The `auth/` group contains the login, register, email confirmation, password forgot/reset, and invite acceptance flows. The `chat/` and `new-chat/` pages manage AI conversation interfaces with message streaming, agent selection, knowledge base attachment (`useKnowledgeBaseAttachment` hook for adding/removing KBs on existing threads, plus pending-KB context state for new threads), and side-panel artifact editing flows such as changing an artifact's assigned letterhead. The `chats/` page lists conversation threads. The `agents/` and `agent/` pages handle AI assistant browsing and detailed configuration. The `prompts/` page provides prompt library management. The `knowledge-bases/` page lists knowledge bases with creation and deletion, while `knowledge-base/` provides the detail view for a single knowledge base including document upload, removal, and property editing. The `skills/` page lists available skills, and `skill/` shows skill details. The academy learning content spans three page slices, one per route: `academy/` lists chapters (`AcademyPage`), `academy-chapter/` shows a single chapter with an embedded Loom video (`ChapterDetailPage`, using the `toLoomEmbedUrl` helper in `lib/` to rewrite Loom share URLs into embeddable iframe URLs), and `academy-chapter-quiz/` runs the chapter quiz (`ChapterQuizPage`); once the whole academy is completed, both the quiz's `QuizResultCard` and the completion banner on `AcademyPage` offer the KI-Führerschein certificate PDF download via the shared `useDownloadCertificate` hook from `features/academy`. The `install/` page handles app installation flows. User `settings/` includes account, general preferences, and `chat-settings/` for per-user AI conversation configuration (custom system prompts). The `admin-settings/` group covers organization administration: model configuration, MCP integrations, team/user management, security, letterhead management (list in `letterheads-settings/`, detail/edit in `letterhead-detail/`), and usage analytics. The `super-admin-settings/` group provides platform-wide management for the models catalog, organization oversight, and platform configuration including per-tier fair-use message budgets via `FairUseLimitsSection` (with `useFairUseLimits` and `useSetFairUseLimit` hooks performing `windowMs ↔ windowHours` conversion at the hook boundary). Each page uses the hooks pattern—one custom hook per operation (create, update, delete)—encapsulating TanStack Query mutations with toast notifications and cache invalidation.
