+++ ayunis-core-frontend/src/pages/academy-chapter-quiz/index.ts
@@ -1,0 +1,1 @@
+export { default as ChapterQuizPage } from './ui/ChapterQuizPage';
