+++ ayunis-core-frontend/src/pages/academy-chapter-quiz/ui/ChapterQuizPage.tsx
