+++ ayunis-core-frontend/src/pages/academy-chapter/index.ts
@@ -1,0 +1,1 @@
+export { default as ChapterDetailPage } from './ui/ChapterDetailPage';
