+++ ayunis-core-frontend/src/pages/academy-chapter/lib/toLoomEmbedUrl.ts
