+++ ayunis-core-frontend/src/pages/academy-chapter/ui/ChapterDetailPage.tsx
