+++ ayunis-core-frontend/src/pages/academy/index.ts
@@ -1,3 +1,1 @@
 export { default as AcademyPage } from './ui/AcademyPage';
-export { default as ChapterDetailPage } from './ui/ChapterDetailPage';
-export { default as ChapterQuizPage } from './ui/ChapterQuizPage';
